Moved cod parsers to cod/
