+++ source/mdcii/mdcii/src/cod/cod_parser.cpp
@@ -29,7 +29,7 @@
 
 #include <google/protobuf/util/json_util.h>
 
-#include "cod_parser.hpp"
+#include "cod/cod_parser.hpp"
 
 Cod_Parser::Cod_Parser(const std::string& cod_file_path, bool decode, bool debug)
   : path(cod_file_path)
