+++ source/mdcii/mdcii/src/menu/gamewindow.cpp
@@ -27,7 +27,7 @@
 #include "sdlgui/widget.h"
 
 #include "bildspeicher_pal8.hpp"
-#include "cod_parser.hpp"
+#include "cod/cod_parser.hpp"
 #include "files.hpp"
 #include "fps.hpp"
 #include "kamera.hpp"
